requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.1.2
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.3
enum34==1.1.6
ipython==5.4.1
ipython-genutils==0.2.0
Markdown==2.6.8
orm-choices==0.2.2
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
